+++ server/client/src/components/Button.jsx
@@ -5,7 +5,7 @@ function Button(props) {
     return(
         <div className="button-container">
               <Link to={props.link}>
-                  <button class="w-25 rounded-0 p-3 btn btn-primary btn-lg">
+                  <button className="w-25 rounded-0 p-3 btn btn-primary btn-lg">
                         {props.ButtonName}
                   </button> 
               </Link>
