+++ server/client/src/components/SubHeader.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 function SubHeader(props){
     return(
             <div>
-                <p class="w-25 text-secondary d-inline-block">{props.SubTitle}</p>
+                <p className="w-25 text-secondary d-inline-block">{props.SubTitle}</p>
             </div>
             )
 }
